+++ app/Http/Controllers/V1/ProduitController.php
@@ -5,42 +5,88 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Produit;
 use App\Http\Requests\Admin\ProduitFormRequest;
+use Illuminate\Routing\Controllers\HasMiddleware;
 use Illuminate\Support\Facades\File;
 
-class ProduitController extends Controller
+use Illuminate\Routing\Controllers\Middleware;
+use Laravel\Sanctum\PersonalAccessToken;
+
+use App\Models\Vente;
+use Illuminate\Support\Facades\Storage;
+
+class ProduitController extends Controller implements HasMiddleware
 {
-    public function index()
+
+   public static function middleware()
     {
-        $produit =Produit::orderBy("created_at", "DESC");
         return [
-            "produits" => $produit,
+            new Middleware('auth:sanctum'),
+        ];
+    }
+
+    public function index(Request $request)
+    {
+        $tokenFromRequest = PersonalAccessToken::findToken($request->bearerToken());
+        
+        $query = Produit::query()->where('user_id', $tokenFromRequest->tokenable_id)->orderBy('created_at', 'desc');
+
+        $produits = $query->get();
+        
+        return [
+            "token" => $request->bearerToken(),
+            "produits" => $produits,
+            "status" => "200", // OK: Indicates that the request has succeeded.
+        ];
+    }
+
+    public function supeAZero(Request $request)
+    {
+        $tokenFromRequest = PersonalAccessToken::findToken($request->bearerToken());
+        
+        // $query = Produit::query()->where('user_id', $tokenFromRequest->tokenable_id)->where('quantite', '>', 0)->orderBy('created_at', 'desc');
+       
+        
+        $produits = Produit::supeAZero($tokenFromRequest->tokenable_id)->get(); //->where('user_id', $tokenFromRequest->tokenable_id)->orderBy('created_at', 'desc')->get();
+        
+        return [
+            "token" => $request->bearerToken(),
+            "produits" => $produits,
+            "status" => "200", // OK: Indicates that the request has succeeded.
         ];
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(ProduitFormRequest $request)
+    public function store(ProduitFormRequest $request): array
     {
-        $fields = $request->validated();
-        $vente = $produit = Produit::create($fields);
+        // dd($request->all());
+        // $fields = $request->validated();
+        // $produit = Produit::create($fields);
 
+        $data = $request->validated();
+        if (request()->hasFile('image')) {
+            
+            
+            $image = $request->file('image');
+            $filename = $image->getClientOriginalName();
+            $imageName = time().'-'.uniqid().'_'.$filename;
+            dd( $imageName);
+            $path = 'pictures/produit/'. $request->user_id;
+           
+            $imageS3 = $image->storeAs($path, $imageName,  's3'); 
+            $url = env('AWS_URL') . '/' . $imageS3;
+            
+            // $imageS3 = env('AWS_URL').'/'.$imageS3;
+            // dd($imageS3);
+            $data ['image'] = $url; // 👈 Make it publicly accessible
+        }
+        $produit = Produit::create($data);
         return [
-            "vente" => $produit,
+            "produit" => $produit,
+            "status" => "200", // OK: Indicates that the request has succeeded.
         ];
-        // $data = $request->validated();
-        // if (request()->hasFile('image')) {
-        //     if($image = $request->file('image')){
-        //         $filename = $image->getClientOriginalName();
-        //         $imageName = time().'-'.uniqid().'_'.$filename;
-        //         $path = 'pictures/produit/';
-        //         $data['image'] = $path.$imageName;
-        //         // $image->storeAs($path, $imageName, 'public'); 
-        //         $image->move($path, $imageName);
-        //     }
-        // }
-        // $produit = Produit::create($data);
-        // // $produit->types()->sync($request->validated('types'));
+        // $produit->types()->sync($request->validated('types'));
         // return to_route('admin.produit.index')->with('success', 'Le produit a été créé');
     }
 
@@ -63,10 +109,22 @@ class ProduitController extends Controller
         if (File::exists($produit->image)) {
             File::delete($produit->image);
         }
-        $produit->update($data);
+         $produit->update($data);
+        return[
+            "produit" => $produit,
+            "status" => "200", // OK: Indicates that the request has succeeded.
+        ];
         // $produit->types()->sync($request->validated('types'));
-        return to_route('admin.produit.index')->with('success', 'Le produit a été modifié');
 
+    }
+
+    public function show(Produit $produit)
+    {
+        
+        return [
+            "produit" => $produit,
+            "status" => "200", // OK: Indicates that the request has succeeded.
+        ];
     }
 
     /**
@@ -74,12 +132,20 @@ class ProduitController extends Controller
      */
     public function destroy(Produit $produit)
     {
+        $vente = Vente::where('produit_id', $produit->id)->first();
+        // dd($vente);
+        if ($vente != null){
+            return ['errMessage' => "Ce produit a une vente"];
+        }
+
         if (File::exists($produit->image)) {
             File::delete($produit->image);
         }
-        //dd($produit->image);
-        //$produit->delete();
-        return to_route('admin.produit.index')->with('success', 'Le produit a été supprimé');
+        $produit->delete();
+        return [
+            "message" => "Le produit a été supprimé",
+            "status" => "200", // OK: Indicates that the request has succeeded.
+        ];
     }
 
     public function destroyImage($produit)
@@ -89,6 +155,9 @@ class ProduitController extends Controller
             File::delete($produitImage->image);
         }
         $produitImage->update(['image'=> '']);
-        return to_route('admin.produit.index')->with('success', 'Le produit a été supprimé');
+        return [
+            "message" => "L'image du produit a été supprimée",
+            "status" => "200", // OK: Indicates that the request has succeeded.
+        ];
     }
 }
